backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-jose[cryptography]
python-dotenv
email-validator
requests
certifi
